requirements.txt
bpy==4.0.0
ftfy==6.2.3
geopy==2.3.0
ipykernel==6.29.5
lxml==4.9.2
mutagen==1.47.0
numpy==1.23.5
openpyxl==3.0.10
pandas==2.2.2
Pillow==10.4.0
pygltflib==1.16.2
PyPDF2==3.0.1
python_magic==0.4.27
PyYAML==6.0.2
Requests==2.32.3
scikit-learn==1.5.1
tabulate==0.9.0
ultralytics==8.2.75
